refactor(foundation-demo): rename header reducer to sidebar reducer
rename header reducer to sidebar reducer
