+++ apps/foundation-demo/src/app/misc/header/actions/sidebar.actions.ts
@@ -1,0 +1,14 @@
+import { Action } from '@ngrx/store';
+
+export enum SidebarActionTypes {
+  ToggleShowSidebarAction = "[Foundation-Sidebar] Toggle Show Sidebar Action"
+}
+
+export class ToggleShowSidebar implements Action {
+  readonly type = SidebarActionTypes.ToggleShowSidebarAction;
+
+  constructor() {}
+}
+
+export type SidebarActions =
+    | ToggleShowSidebar;
